EN-5664: Limit crawling of the catalog with robots.txt.
Currently 40% of our traffic is from Bing, mostly crawling different
combinations of the catalog.  That is silly.  A better solution
would be nofollow links on the individual pages, but this is a lower
risk easier approach that will hopefully address the issue.

We are explicitly trying to allow a single parameter because we
have cases like https://data.seattle.gov/browse?category=Public+Safety
which currently do have good SEO (top result for "seattle public
safety data" on Google while limiting the cases that cause trouble.

Also, add a crawl delay setting as backup.