EN-5711: Add more catalog pages and misc. to robots.txt.
Turns out we expose filtered catalog pages at a variety of URLs,
so attempt to block more of them.

Also add /login and */edit$ which we are seeing a reasonable amount
of bot traffic to (especially, but not limited to, GSA) and which
don't seem useful fot bots to index.