EN-5453: Don't wait for all callbacks before rendering Discuss pane
This is a sad, sad workaround for the fact that fetching related views
can take an obscenely long time on particular datasets, which causes
the Discuss pane to effectively timeout in rendering.

The real fix is to make the relevant Core call for related views more
performant, since the same problem unavoidable impacts More Views.