ENG: Remove calls to route_params where possible
This restores the patch to ActionDispatch URL helpers that handles automatically calling route_params when the first argument to the helper is a Model class that implements the route_params method. This is the way the models worked in Rails 3 and we initially didn't want to port the patch to Rails 4. This commit represents the fact that we changed our minds and actually did want to port the patch.

Signed-off-by: Randy Antler <randy.antler@socrata.com>