EN-4053: L02 (Legacy) Use simplified presets (#2650)
Simplifies the "reset to preset" option by only offering
one option "restore suggested settings".  Also updates the
default to separate discrete and composite columns.  Note that this
explicitly excludes the AC of disabling the button when the settings have
been restored as I was unable to easily isolate a single place where
we could re-enable the button once the columns changed again.