BUG: Remove feature flag hacks since it was dalli / memcache
Turns out that since we use nutcracker/twemproxy for HA with memcached, which incidentally doesn't support the memcached binary protocol; and since Dalli uses the memcached binary protocol; and since modern Rails apps _refuse_ to let us use a memcache client from 2010, we have to figure out a new HA story for memcached. In the mean time, the work around it to change the chef cookbook to write the address of one of the memcached nodes into the production.rb file instead of the nutcracker address.

Signed-off-by: Randy Antler <randy.antler@socrata.com>