EN-3608 Allow users to share datasets with groups
Users could share their  datasets with other users or any arbitrary
email. This patch adds the ability to share datasets with groups.

In order to share a dataset with a group, the user clicks  Share Dataset
button under Manage. As the user starts typing, autocomplete box will
appear to suggest user to his/her contacts and groups in that domain. If
the user selects a group from the list, a group grant will be created to
give permission to the group over the dataset.

To implement that:
  1 - An AJAX request to fetch groups for autocomplete added
  2 - A hidden input added to HTML to store group id
  3 - Autocomplete is now aware of the type of selected element
      user or group
  4 - New validation method emailRecipient added to validate input
  5 - Grant object will be constructed according to type of the
      autocomplete selection