EN-6382 - Handle invalid UTF-8 in URL parameters
Because of <String>.encode being non-mutating, although we were forcing UTF-8 encoding and replacing invalid byte sequences with an empty string we were never actually reassigning each param with the newly encoded and replaced string, so the params that the rest of the application was using were the original, unmodified ones. The modified ones were presumably lost in lambda-space.

* Add the ! in o.encode! so that we actually force UTF-8 and replace invalid byte sequences with empty strings for each param in the params hash.