EN-6234: Fix not being able to scroll when viewport set to device-width
If there is a viewport tag with `content="width=device-width"` set, we
were unintentionally adding `overflow: hidden` to the body by calling
`positionFlannel` after intitializing the flannel when in mobile. We
had added that to prevent the flannels from making the page extra long,
but since we're setting using `display: none` instead of
`visibility: hidden` for `.flannel-hidden`, that shouldn't be an issue.