fix unicorn to listen on 3000, unless otherwise specified (#2798)
The docker container is running nginx as a side car process listening on port 3000. We now set an environment variable in the Dockerfile to make unicorn listen on 3333. This reverts the current behavior which broke frontend locally, and special cases the behavior we require for frontend running within a docker container.