EN-7517 - Temporarily override viz title and description with null
The new viz types actually read from title and description and will display them, but the VisualizationAdd controller will set the title to the name of the column or something.

Until users have a way to actually change the title and description, we want to make sure that we override them at runtime to null, which will prevent them from being displayed.