EN-4551: Makes html embed XHR requests unique to walk around caching;
- Adds a query param called from_xhr to html embed:

Why?

from_xhr walks around browser caching, which was pulling an earlier
response from the <iframe> we load with the HTML fragment. The difference
in headers from the requests causes this XHR to fail due to S3 deciding whether
or not to provide access control headers based on an "Origin" request header.
This header is not provided through an <iframe> HTTP request like it is
through our XHR, so when the cached request and response are derived from the
<iframe> request, the browser/jQuery become unhappy. Adding a query param
creates a distinct request that can't be cached by association.