EN-1673: Stop using Memcached for DataSlate and see what breaks. (#2806)
If that sounds scary, that is because it is scary.

Feature Flag name: `route_dataslate_without_caching`.
Valid values: true, false.

A DataslateRouting object now handles routing for Dataslate pages.
Pages are not meant to be meaningfully cached (there is a little
in-process caching). We no longer scrape the entire table for pages in
order to route them; instead, we only request the full content when we
want to render them.