EN-7231: adds a show page for data connectors (#3023)
* add link to connectors index page on # assets connected
 * show page links to arcGis layers and socrata datasets
 * pagination
 * additionally, start sending posts to esri crawler with domain id

QA: on a domain with the feature flag 'enable_catalog_connector' turned on, go to admin/connectors
For a connection that has been added, with at least one layer selected to connect to, the '# of assets connected' should be a link to /admin/connectors/:id
On that page, there should be an entry in the table for every connected layer. The links on the right should go to socrata external datasets. On the top right should be a status of the server, including the time of the last successful sync.