EN-7861: Fix grants when sharing a dataset
When the sharing pane gets the grants, they usually have ONE OF an email
or a user id. However, in some cases, a grant might habe BOTH an email
and a user id.

When changing/removing a grant, the object sent to core has to EXACTLY
match what core sees as the grant object; this means email, user id, or
both.

This makes it so that if both are present, they properly get sent.