EN-1313: Maybe fix unsaved maps; (#3086)
Thoughts: The unsaved thing is showing up when you don't have update privileges because
the 'setQuery' function gets called a lot during initial loads of maps and this is always
considered a "change" on the dataset, which marks it as temporary which shows the "unsaved view"
message.  We seem to have a concept of `_ignoreTemporary` in maps which prevents setting of the
temporary flag.  Maybe we should also use it in `setQuery`.  Preliminary testing confirms that
maps no longer load as unsaved but correctly become unsaved if you pan/zoom.