EN-6353: Integrate styleguide into catalog/browse2 (#3117)
This is a a first-pass effort only which integrates the inclusion of the styleguide onto the catalog page and addresses any styling issues that arose as a result. The DOM structure of the existing catalog output has not been modified to match the catalog as defined in the styleguide. This will be addressed when we convert the catalog into a reusable component.