EN-9292: Styleguide flyout bindings can't bind to something missing;
We conditionally render the flyout based on the data that comes back
from our view API. If there isn't enough, we don't render the flyout,
but the Styleguide still attempts to bind to whatever data-flyout
specifies. To resolve this, I'm only adding data-flyout when we actually
intend to render the flyout.