EN-6784: fix a race condition in how useSODA2 + column .lookup are set (#3244)
* EN-6784: fix a race condition in how useSODA2 + column .lookup are set

Found with the help of Michael Chui! Basically what was going on here
was NBE datasets were:
1) doing column lookup setting twice (once to int id and once to fieldname)
2) sometimes calling things like soql_clustered in between the two
lookup sets (which would fail because int ids aren't valid for that
call), and sometimes calling it after the second lookup set (which
worked normally)
So the net effect of this should be removing some intermittent errors
with soql_clustered (and possibly other functions as well).

* Add comment to dataset.js so we know why this code is called where it is.

In the words of Chui: What's going on in this change is that the
functionality for setting _useSODA2 is being moved into its own function
and getting called directly by #_init before columns are constructed for
the dataset, because it turns out there's a thing in column.js that
actually pays attention to that property... which... isn't... set at the
time of creation.