EN-7855: Stops "Try again" save button from spinning into oblivion;
I know we all really enjoy putting our "Try again" button text in the
windowed washer and watching it spin, but it's time to end the fun.

When you click to reattempt to save the story after an error, the "Try
again" button correctly disappears and a spinner replaces it.