EN-8903: DSLP fix tour position when titles wrap (#3339)
The little checkmark icon next to "official" in the info pane can load
after the tour initializes, which sometimes caused the info pane buttons
to move to the next line. That resulted in the tour arrow pointing where
the button was before the icon loaded. To fix it, let's make the icon
element take up as much space as it would when the icon loads so we
don't have things shifting unexpectedly when the icon finishes loading.