EN-8981: Fix viewport for public maps derived from a private ESRI-connected map (#3344)
If you try to access a public map that is derived from a private
ESRI-connected map that you don't have permission to access, the map
currently loads zoomed all the way out. This was happening because we
try to start loading the map with the viewport set in the parent view's
metadata. If you don't have permission to view the parent view, we try
to figure out the bounds from undefined and then everything falls apart.
If the parent view is undefined, let's not do that.