EN-9776: Loads orderBy into AX; (#366)
Turns out we forgot to update the loader when we added the new field to
AX's configuration. We've done this on several occasions now. It might
be useful to have some documentation about adding new configuration
fields to AX and what steps you need to take to make that successful.

At this point, we've got it down to a simple workflow.