EN-9791: Add goToPreviousPage function to enable navigating backwards better
Make error navigation easier by providing a `goToPreviousPage` function
that cleans up pages on the path as it returns a user back to an
arbitrary page. This is important for cases where we are making the
navigation calls in a async fashion, and we aren’t completely sure
which page the user is currently on.