EN-6820 Use cetera / new catalog for all new domains (#3123)
* ENG: OMG! Such campsite cleansing!

* ENG: Make it easier to spoof Cetera in development

Setting the environment variable `CETERA_SPOOF_HOST` will cause Cetera to use this host for the `domains` and `search_context` parameters when making queries.

* EN-6820: Use cetera/new catalog on new domains

Instead of relying on `view_type` to decide if the new catalog should be used, we now use the feature flag `cetera_search` feature flag, since the use of Cetera _requires_ the new catalog.

Included in this PR is a change to the `stub_feature_flags_with` and `rspec_stub_feature_flags_with` helper methods. They no longer take seperate key and value args, and instead take a hash with one or more entries. Multiple calls to the helper methods merge the stubbed feature flags.

* ENG: Fix flaky tests.

* ENG: Fix bad merge in site_chrome translations

* LINT: Remove redundant parenthesis

* ENG: Fix bad merge with stub_feature_flags_with

* ENG: Style

* ENG: Rename to conform to naming conventions

* ENG: Remove deprecated tests

* Remove SCSS loader

* TEST: Fix breaking tests