EN-10412: Resolve dsg bug re: eslint changes (#3445)
This is a re-implementation of eslint fixes that resolves an issue mistakenly introduced by the first eslint fixes.

The symptom was that core requests were made to the row service that didn't include the full data of the view.
(cherry picked from commit b511397c3a8933b601a7efc0f12687368afcd46a)