EN-10022: Allows selection of duplicate picklist values.
When options contained multiple values in the past, keyboard navigation
would reset to the first item in the list. The navigation now acts as
you'd expect, by continuing to the next item in the options list. We
also render whatever number of selected options that have the same
value in the same highlighted state.