EN-9317: Keeps table form preventing shift tab; (#3542)
When trying to shift+tab out of the table, it forces focus on the first
table cell. That means a keyboard user cannot escape the clutches of the
table when they want to. This work ensures that if there isn't any
navigation, we properly allow the user to shift+tab to the previously
tab-indexable item on the page.