EN-9428: Fix Choropleths in Polaroid (#3577)
Choropleth needs the fieldName value to exist in the columns it's using
in order to render. As it turns out, we were decorating the columns with
the fieldName in the table controller, which meant that choropleths
could only render if there also happened to be a table on the page,
which doesn't really work when we're rendering a choropleth by itself
for Polaroid.

This change moves the fieldName decoration to the dataset model, so
everyone can use it.