EN-9560: Remove displayFormat.forceBasemap
This was deprecated when we upgraded maps to OpenLayers in 2011. There
are no traces of it in any of the metadbs and the only references to it
that I can find are both in conditional statements in frontend. I'm
keeping it as a separate commit *just in case*, but I think this is
pretty safe to remove.