Rantler/en 11117/unicorn stats gathering (#3586)
* EN-11117: Add UnicornMetrics middleware

This adds middleware to allow us to see usage metrics for the Unicorn workers in the frontend.

The access path is currently configured to be "/metrics", but this may be changed via the ENV variable UNICORN_METRICS_PATH.

The URL path should protected from public access. That change will be addressed in a separate PR.

* ENG: Whitespace and explicitly set preload_app

Documenting the default state of preload_app as false. Storyteller uses true and we may want to do that as well in the future.

From the Unicorn docs:
"The preload_app directive loads the application before forking workers, so it can share any loaded data structures. By default, workers each load a private copy of their app for out-of-the-box compatibility with existing servers."

"preload_app can dramatically speed up startup times. It can also make it easy to share memory across processes when using Ruby Enterprise Edition. REE also uses tcmalloc on some platforms, like Linux, instead of a generic malloc, which improves performance for most server workloads independently of copy-on-write."

Ruby Enterprise Edition was EOL in 2012, so I don't think that comment in the cos is relevant any longer.

* PR: Address PR feedback

* GEM: Bump gem version