EN-11152: Fixes dropdown scroll event handling; (#180)
In some cases we were trying to bind to things that weren't necessarily
scrolling and things that weren't necessarily HTMLElements. This causes
odd binding issues later in the React component's lifecycle that are
only made obvious in host applications like Storyteller.

This fix makes the bind logic stricter by requiring overflow and a
distinguishable scroll height difference from the client height. There
are cases where the difference is by a pixel and there's not a
scrollbar, this should be covered by the overflow check.