EN-11152: Stops null thrashing when rendering recommended flyouts; (#401)
When the Authoring Workflow is embedded in Storyteller, a series of
random errors would pop up while navigating the dimension picklist.
Turns out that the picklist was causing a series of re-renders across
the entire application. While that may seem innocuous, the flyouts that
toggle for recommended visualization types didn't think so. The logic
constantly swaps between a null-rendered state and element-rendered
state. This makes React super unhappy. To save re-rendering cycles,
the recommended labeling now relies on `classNames` and its parent for
visibility.