EN-11152: Simplifies pie chart layouts; (#405)
Let's not worry about cutting the pie into two different layouts. When
we do that, the width < height case doesn't scale nicely. Instead, we
always try to maximize the width of the pie chart based on height up to
some maximum, clamped value. This fills out the container as one would
expect. The sizing and layout for legends remains.

There is a case when the height becomes so small that the legend (in a
width > height calculation), is cut off by the container. We've accepted
that tradeoff for the other cases.