EN-11223: Fix SDP OData link; (#3585)
Currently, there is an "API" checkbox on the SDP template page
that toggles the api and the odata controls on the SDP preview.
However, on the actual SDP page, the OData control always shows
up.  The hypothesis is that this is gated on an old "odata"
checkbox that no longer exists.