EN-11358 Improve pieChart label visibility
Instead of filtering pie slice labels according to angle difference now
they are shown and hide according to overall chart size. It computes
visible area for given pie slice and hides labels if there is not enough
space or shows if there is enough space. This functionality called in
resize handler, so showing/hiding labels will be happening live during
the resize action.