EN-11631: Adds peerDependencies and upgrades node to 6.9.1; (#187)
The days of including things from a CDN are over for Styleguide. We
moved to the frontend-visualizations model of building a separate
vendor.js based on Webpack externals. This eliminates the need to copy
files around using Middleman, worrying about CDNs being unavailable, and
maintaining too many lists of vendor software. Things are neatly placed
in package.json, webpack/, and pages/javascfripts/vendor.js.

This update brings the requirement of peerDependencies and their
management in host applications. That means you can't install Styleguide
without matching our required versions. This means we're no longer
bundling multiple versions of React (or any other library for that
matter) that is also used in the host application.