EN-11631: Adds peerDependencies and upgrades to React 15; (#417)
Things to note:
- Debounced events were restructured and abstracted to remove concerns
  around React's synthetic event system.
- We've added Styleguide's peerDependencies requirements.
- EN-9281 has been removed from AX until the feature is completed.
- Webpack externals were updated.
- React was upgraded to 15 to match frontend.

This work successfully nets us the reduction in bundle size. With the
upgrade to node v6, we began bundling React (among several others) twice, but with
peerDependencies, the host application and this library will now be
required to match versions. This requirement helps Webpack optimize module
bundling.