EN-12080 Fix dropdown at bottom doesnt scroll
If the dropdown located at the bottom of the page it doesn't scroll.
After rendering the component it calculates the maximum height of the
dropdown list, but mistakenly dropdownlist container and the picklist
component's height set together, that causes wrong calculation of
scrollHeight on next render since container and the child has the same
height scrollHeight points wrong height.

This commit remove the lines which sets the picker's height and adds
scrolling to picklist's container.