EN-12328: Lazily initialize site_chrome instance (#155)
* EN-12328: Lazily initialize site_chrome instance

This changes how we initialize the site_chrome instance in the middleware. Before we would always instantiate it and set the instance as a property on Rails.application.config. This caused the site_chrome to be instantiated for every request, including the version and health check requests. In frontend, the middleware are ordered such that the version middleware comes before the site_chrome middleware. In StoryTeller this wasn't the case. We _could_ have resolved this by using middleware in StoryTeller to provide the health check, but we instead chose to change SiteChrome to use lazy initialization, since this is a more generalized solution to the problem.

* PR: Address code review comments.

* Move site_chrome_instance to shared helpers module