EN-12399: Block crawling of alt pages to reduce load.
The specific thing that lead here was metadb getting pegged from
getByTableId calls for NYC 311.  The real fix for this is to move
getByTableId to use cetera or a cheaper getByTableId that doesn't
return all the info, but looking into this further revealed that
some bots are crawling datasets page by page through the alt page.
This puts a heavy load on the system.  Google doesn't seem to index
alt pages and instead prefers export, so the bang for the buck from
remaining crawlers isn't very high.  Also, ending up on the alt
page isn't a great experience anyway.

Once getByTableId is fixed it may be worth considering allowing
crawling the first x (say... 5) pages of the alt page... maybe.