EN-12385: Fix social link page js (#3812)
Previously, this javascript would switch the login and sign up forms but now that we have a new login page, we just want to redirect. The javascript was removed so that the "SIgn In" link in the top right would just redirect, but apparently the javascript was a little overloaded.