EN-13042: Renders custom_css based on user's login. (#3915)
This fix moves us over to the suppress_govstat? helper function that
checks the logged-in state of the user to determine if custom CSS should
be loaded. Before, we were relying on the instance variables for
suppress_govstat which was commonly "nil". An unset variable is a
iffy indicator whether or not we should load the CSS; better to have an
explict acknowledgement of true or false, which is what
suppress_govstat? provides.