EN-11687 - Grouped Multi-series Bar and Column Charts
The original intent, which was to map each unique value in the dimension to a series at chart creation time, was not acceptable because changing data might have resulted in invalid vifs.

Instead, we chose to dynamically generate a 'fake' vif representing the unique dimension values at render-time. We do this by making many individual queries and then stitching the results together into a single data table.

This resulted in a few refactors of how bar and column charts have historically handled the 'multi series' case:
- They both now use the CategoricalDataManager, which has abstracted away data fetching for categorical columns into a single module that both use.
- They also both now expect a single data table with multiple measure columns, as opposed to an array of data tables with one measure column each. This approach appears to be more consistent with how d3 represents data internally and allowed us to simplify the implementations of charts themselves.

Many smaller changes were made, including a few fixes to tests that were fundamentally broken and a fix to the positioning of the flyout when highlighting a group of measure values that might fall outside the defined measure axis min and max values.