EN-13299: Address Airbrake spam from jQuery AJAX in IE11
The lesson here is that Airbrake messages should be more or less consistent,
and putting parameters or arbitrary data into the message is not a good idea.

This change led to a cleanup pass on httpRequest, getting rid of some older
methods of making AJAX calls as well as changing the signature for the
success resolver of httpRequest itself.