EN-13218: Don't show dervied views on alt pages for performance reasons.
The intent is to bring this back at some point, pulled from cetera,
but that has been delayed for some time so we want to stem the
bleeding.

This is the cause of the last three site outages due to the expense
of this on the core/metadb side, see the related ticket for more
details.  We already block "real" bots from this in robots.txt but
see ad-hoc bots crawling large numbers of pages.

FUN FACT: This actually calls getByTableId twice, once just for the
"@view.filters.size > 0" check.