EN-9218: Move form error rendering to core
This removes blindly rendering the error message in the query string
redirected from core.  Changes are being made in core to directly render
the error page to remove the opportunity for hijacking the displayed error
message.

A follow-up changeset will address completely removing the error page
and passing the error page redirect.