EN-14046: Stops jarring scroll after filter selection. (#260)
* EN-14046: Stops jarring scroll after filter selection.

When selection an option from "Add Filter", the picklist is forcibly
focused. To walk around this situation instead of fixing it, this PR
resizes the searchable picklist to the height of the browser's bottom
edge.

This problem doesn't seem to show up with the TextFilter component,
oddly.

* En-14046: Revert auto height implementation. Introduce delayed focus.

The idea is to delay the focus, giving the parent component time to do
what it wants with the focus, before trying to focus the picklist.