EN-14579: Allow search card mapping but do not bootstrap it (#4206)
* EN-14579: Avoid bootstrapping search cards

The original implementation in EN-13836 was to make the _default_ card type mapping for search cards “invalid” to avoid them getting bootstrapped. However, the DataLens implementation currently prevents adding a card if the default card type is “invalid”, _despite_ the fact that _available_ card types are present. This change was the smallest delta to achieve the desired outcome without wading into the ancient and difficult Angular code of DataLens.

* EN-14579: Avoid bootstrapping search cards

The original implementation in EN-13836 was to make the _default_ card type mapping for search cards “invalid” to avoid them getting bootstrapped. However, the DataLens implementation currently prevents adding a card if the default card type is “invalid”, _despite_ the fact that _available_ card types are present. This change was the smallest delta to achieve the desired outcome without wading into the ancient and difficult Angular code of DataLens.

* EN-14579: Reject search cars at the last moment

Man this controller and the test are huge hot mess. The whole thing needs a massive refactor.

The original logic for EN-13836 was to avoid bootstrapping search cards by changing the default_card_type_for method to make them ‘invalid’, however DataLens code in AngularJS examines the default_card_type property rather than availableCardTypes when determining if a column can be added to a card. This resulted in making the search card unavailable for high cardinality columns (why it remained available for low cardinality columns is left as an exercise for the reader).

The new logic simply waits until the very last possible moment during the (extremely circuitous) bootstrapping process to elide any search cards from the page metadata and makes no changes to default_cardType at all.

(cherry picked from commit 3b44bbd8e4bbf2cf16c6f7bec4706af430777ed1)

# Conflicts:
#	test/unit/card_type_mapping_test.rb