EN-14735: Fixup frontend logging to have severity, better timestamps, etc.
This rejiggers things to include the severity in log messages and log the time of the actual log
call and not use the request start time for every log call.

It also updates the unicorn logging to be more consistent and the RequestLoggerMiddleware logging
to be in a more consistent form (it already wasn't CLF...), include useful things like request id,
remove things we can't log, and actually log the query string.