EN-14047: Restructures map zoom and center AX actions. (#548)
We no longer request and save zoom and center. This is critical to avoid
a circular action. Now, we only re-render our VisualizationPreview if
the visualization has anything but a zoom and center change. Why? Zoom
and center are mutated internally by the Feature Map and Region Map.
That means we shouldn't feed the information back into the map because
it has no data diffing mechanism to save us from an infinite loop.

This problem will crop up with an internal state that AX is interested
in persisting in its representations of each VIF.