EN-14706: Removes SoQL query trim. (#549)
* EN-14706: Removes SoQL query trim.

Alright, so here's the deal: we roll with multiline-base query strings
in our SoqlDataProvider which ends up adding a bunch whitespace and
newlines and carriage returns and walruses (but not really walruses).
Someone back in the day thought it would be a Good Deed™ to remove all
of that extra stuff to make queries look nicer and fight the query
parameters character limit. The thing is, whitespace-dependent values
get completely obliterated.

For example:

select `yes` from anywhere where string = 'Holy FANTASTIC      ';
// is converted to
select `yes` from anywhere where string = 'Holy FANTASTIC ';

That makes for one angry walrus, so let's just not do that. There are
several ways to fix this, but really, we should no longer be using GET
for our queries. We should move to a POST system.

I'll be writing another ticket to bring this up.

* EN-14706: Exchanges multiline template strings for arrays.

I'm beginning to suspect that multiline template strings should not be
used for code legibility as they have real consequences on underlying
data.

* EN-14706: Improves localhost examples support.