EN-13453: Handle hidden columns in requests (#557)
As it turns out, if you are logged in and have access, the request for
metadata will return all of the columns, with the hidden columns
decorated with a nice "hidden" flag. Our code previously assumed that
hidden columns would never be returned with the metadata.

This bug manifested when looking at table visualizations as a logged
in user. The hidden column title would continue to be displayed, but
there would be not data included in the column and the data in all
subsequent columns would be off by one.