EN-15425: Fix adminGoals linting error (#4390)
adminGoals should be using the same linting rules as the other React
projects. Since that update will involve fixing 700+ linting errors,
we're disabling the offending rule for the time-being. The rule is to
enforce not using dot notation for reserved words, which is an
irrelevant issue now that we're using webpack instead of jammit
everywhere.

We'll update the linting rules in EN-15427.