EN-14630: Draft assets link to DSMP in the profile if you have edit rights
* specifies current_revision path

A draft represents a dataset in an unfinished state. When the creation workflow is complete, it is updated to display_type tabular. If you are accessing an incomplete dataset, and have permissions to edit it, you should be taken to the Dataset Management Page

QA: create a dataset in the beta experience. Confirm at /api/views that it has display_type draft. Confirm that the url in the profile points to the full dataset path, ending in /revisions/current.