EN-15800: Fix flashing region maps (#574)
When re-rendering a region map with a new center and zoom, we can get
into an infinite loop where we render at the new center and zoom and
then flash back to the original center and zoom (repeat infinitely). It
turns out it was caused by looking at the original vif instead of
looking at the new vif when re-centering ourselves (because we didn't
update the internal vif we reference until after the center and zoom
update has happened). Whoops.