EN-14758: Publishing story makes copy of attachments
Created a model for working with story blocks JSON and performed a
small refactor. During review of the approach, we explored the option
of refactoring the new StoryJsonBlocks class into (Draft|Published)Story
itself, but that refactor proved to be a bit hairier than we were
willing to take on at this moment.