EN-12778: Upgrade moment and use friendly timezones for m-w message.
There's some version leapfrogging here. What happened?
- We were running on Moment 2.4.0 for a long while.
- Commit dcfe61040e pushed all-screens onto 2.14.1 to get moment.locale
  - This was likely not intentional, but who knows.
- This commit uses Moment 2.18.1.

Why?

Because 2.14.1 is included as `devDependencies` in package.json, and
thus is being included in a frighteningly unreliable manner. Hopefully a
future commit will resolve this issue indepedently. Since I'm upgrading
moment anyways, there's no good reason not to go straight to the latest
version (2.18.1), rather than inching along to 2.6.0 (which is what
moment-timezone requires) or 2.14.1 (which is what we started using only
a release or two ago).